+++ src/main/java/net/nicotfpn/alientech/entropy/EntropyTransaction.java
@@ -4,37 +4,32 @@ package net.nicotfpn.alientech.entropy;
  * Transaction-safe entropy transfer result.
  * <p>
  * Used to ensure entropy transfers are atomic: either fully succeed or fully
- * fail.
- * Prevents entropy duplication or loss from partial transfers.
+ * fail. Prevents entropy duplication or loss from partial transfers.
+ * All amounts are {@code long} to match the upgraded {@link IEntropyHandler}
+ * API.
  */
 public final class EntropyTransaction {
 
-    private final int amount;
+    private final long amount;
     private final boolean committed;
 
-    private EntropyTransaction(int amount, boolean committed) {
+    private EntropyTransaction(long amount, boolean committed) {
         this.amount = amount;
         this.committed = committed;
     }
 
-    /**
-     * Create a committed transaction (transfer succeeded).
-     */
-    public static EntropyTransaction committed(int amount) {
+    public static EntropyTransaction committed(long amount) {
         return new EntropyTransaction(amount, true);
     }
 
-    /**
-     * Create a failed transaction (transfer did not occur).
-     */
     public static EntropyTransaction failed() {
-        return new EntropyTransaction(0, false);
+        return new EntropyTransaction(0L, false);
     }
 
     /**
      * @return the amount transferred (only valid if committed)
      */
-    public int getAmount() {
+    public long getAmount() {
         return amount;
     }
 
@@ -53,14 +48,14 @@ public final class EntropyTransaction {
      * 2. Simulate insertion into destination
      * 3. If both succeed, execute both operations
      * 4. Return transaction result
-     * 
+     *
      * @param source    the source entropy handler
      * @param dest      the destination entropy handler
      * @param maxAmount maximum amount to transfer
      * @return transaction result
      */
-    public static EntropyTransaction transfer(IEntropyHandler source, IEntropyHandler dest, int maxAmount) {
-        if (source == null || dest == null || maxAmount <= 0) {
+    public static EntropyTransaction transfer(IEntropyHandler source, IEntropyHandler dest, long maxAmount) {
+        if (source == null || dest == null || maxAmount <= 0L) {
             return failed();
         }
 
@@ -69,40 +64,35 @@ public final class EntropyTransaction {
         }
 
         // Step 1: Simulate extraction
-        int available = source.extractEntropy(maxAmount, true);
-        if (available <= 0) {
+        long available = source.extractEntropy(maxAmount, true);
+        if (available <= 0L) {
             return failed();
         }
 
         // Step 2: Simulate insertion
-        int accepted = dest.insertEntropy(available, true);
-        if (accepted <= 0) {
+        long accepted = dest.insertEntropy(available, true);
+        if (accepted <= 0L) {
             return failed();
         }
 
         // Step 3: Determine actual transfer amount
-        int toTransfer = Math.min(available, accepted);
+        long toTransfer = Math.min(available, accepted);
 
         // Step 4: Execute both operations atomically
-        int extracted = source.extractEntropy(toTransfer, false);
-        if (extracted <= 0) {
-            // Extraction failed - abort transaction
+        long extracted = source.extractEntropy(toTransfer, false);
+        if (extracted <= 0L) {
             return failed();
         }
 
-        int inserted = dest.insertEntropy(extracted, false);
+        long inserted = dest.insertEntropy(extracted, false);
         if (inserted != extracted) {
-            // Insertion partially or fully failed — roll back by returning entropy to
-            // source
-            int uninserted = extracted - inserted;
-            if (uninserted > 0) {
+            long uninserted = extracted - inserted;
+            if (uninserted > 0L) {
                 source.insertEntropy(uninserted, false);
             }
-            // If nothing was inserted at all, it's a full rollback
-            if (inserted <= 0) {
+            if (inserted <= 0L) {
                 return failed();
             }
-            // Partial transfer succeeded
             return committed(inserted);
         }
